requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
decorator==5.1.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
ipdb==0.13.9
ipython==7.29.0
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
orjson==3.6.4
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pycep-correios==5.0.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.4
starlette==0.16.0
toml==0.10.2
traitlets==5.1.1
typing-extensions==4.0.0
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
validate-docbr==1.10.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
zeep==4.1.0
